Revert "Support grpc-web"
This reverts commit d58659507c0de38de596411ca2890050b6d525f8.

Development of grpc-web will continue on the branch
https://github.com/grpc/grpc-dart/tree/grpc-web until it is ready.
