+++ CHANGELOG.md
@@ -1,8 +1,3 @@
-## 1.1.0
-
-* Add initial support for grpc-web.
-  See `example/grpc-web` for an example of this working.
-
 ## 1.0.1
 
 * Add `service_api.dart` that only contains the minimal imports needed by the code generated by
