--- build.yaml
@@ -1,5 +1,0 @@
-targets:
-  $default:
-    sources:
-      exclude:
-        - example/**
