--- example/grpc-web/protos/echo.proto
@@ -1,41 +1,0 @@
-// Copyright 2018 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-syntax = "proto3";
-
-package grpc.gateway.testing;
-
-message EchoRequest {
-  string message = 1;
-}
-
-message EchoResponse {
-  string message = 1;
-}
-
-message ServerStreamingEchoRequest {
-  string message = 1;
-  int32 message_count = 2;
-  int32 message_interval = 3;
-}
-
-message ServerStreamingEchoResponse {
-  string message = 1;
-}
-
-service EchoService {
-  rpc Echo(EchoRequest) returns (EchoResponse);
-  rpc ServerStreamingEcho(ServerStreamingEchoRequest)
-  returns (stream ServerStreamingEchoResponse);
-}
