--- example/grpc-web/pubspec.yaml
@@ -1,15 +1,0 @@
-name: grpc_web
-description: Dart gRPC-Web sample client
-homepage: https://github.com/dart-lang/grpc-dart
-
-environment:
-  sdk: '>=2.0.0 <3.0.0'
-
-dependencies:
-  grpc:
-    path: ../../
-  protobuf: ^0.13.4
-
-dev_dependencies:
-  build_runner: ^0.10.0
-  build_web_compilers: '^0.4.1'
