--- example/grpc-web/web/index.html
@@ -1,42 +1,0 @@
-<!--
-Copyright (c) 2018, the gRPC project authors. Please see the AUTHORS file
-for details. All rights reserved.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-  http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
--->
-
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Echo Example</title>
-  <link rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
-  <script defer src="main.dart.js"></script>
-</head>
-<body>
-  <div class="container">
-    <div class="row" id="first">
-      <div class="form-group">
-        <div class="input-group">
-          <input type="text" class="form-control" id="msg">
-          <span class="input-group-btn">
-            <button class="btn btn-primary" type="button" id="send">Send
-            </button>
-          </span>
-        </div>
-        <p class="help-block">Example: "Hello", "4 Hello"</p>
-      </div>
-    </div>
-  </div>
-</body>
-</html>
