+++ example/helloworld/pubspec.yaml
@@ -13,4 +13,3 @@ dependencies:
 
 dev_dependencies:
   test: ^1.3.0
-
