+++ interop/lib/src/client.dart
@@ -19,9 +19,9 @@ import 'dart:typed_data';
 
 import 'package:collection/collection.dart';
 import 'package:grpc/grpc.dart';
-import 'generated/empty.pb.dart';
-import 'generated/messages.pb.dart';
-import 'generated/test.pbgrpc.dart';
+import 'package:interop/src/generated/empty.pb.dart';
+import 'package:interop/src/generated/messages.pb.dart';
+import 'package:interop/src/generated/test.pbgrpc.dart';
 
 const _headerEchoKey = 'x-grpc-test-echo-initial';
 const _headerEchoData = 'test_initial_metadata_value';
@@ -95,7 +95,7 @@ class Tester {
       if (_useTestCA) {
         trustedRoot = new File('ca.pem').readAsBytesSync();
       }
-      credentials = new Http2ChannelCredentials.secure(
+      credentials = new ChannelCredentials.secure(
           certificates: trustedRoot, authority: serverHostOverride);
     } else {
       credentials = const ChannelCredentials.insecure();
