--- lib/src/client/http2_channel.dart
@@ -1,33 +1,0 @@
-// Copyright (c) 2019, the gRPC project authors. Please see the AUTHORS file
-// for details. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import 'package:meta/meta.dart';
-
-import 'channel.dart' as channel;
-import 'options.dart';
-import 'transport/transport.dart';
-import 'transport/http2_transport.dart';
-
-@visibleForTesting
-Future<Transport> connectTransport(
-    String host, int port, ChannelOptions options) async {
-  return Http2Transport(host, port, options)..connect();
-}
-
-class ClientChannel extends channel.ClientChannel {
-  ClientChannel(String host,
-      {int port = 443, ChannelOptions options = const ChannelOptions()})
-      : super(host, connectTransport, port: port, options: options);
-}
