--- lib/src/client/web_channel.dart
@@ -1,32 +1,0 @@
-// Copyright (c) 2019, the gRPC project authors. Please see the AUTHORS file
-// for details. All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-import 'package:grpc/src/client/options.dart';
-import 'package:grpc/src/client/transport/transport.dart';
-import 'package:grpc/src/client/transport/xhr_transport.dart';
-import 'package:meta/meta.dart';
-
-import 'channel.dart';
-
-@visibleForTesting
-Future<Transport> connectXhrTransport(
-    String host, int port, ChannelOptions _) async {
-  return XhrTransport(host, port)..connect();
-}
-
-class GrpcWebClientChannel extends ClientChannel {
-  GrpcWebClientChannel.xhr(String host,
-      {int port = 443}) : super(host, connectXhrTransport, port: port);
-}
