+++ lib/src/shared/streams.dart
@@ -20,8 +20,39 @@ import 'dart:typed_data';
 
 import 'package:http2/transport.dart';
 
-import 'message.dart';
 import 'status.dart';
+
+abstract class GrpcMessage {}
+
+class GrpcMetadata extends GrpcMessage {
+  final Map<String, String> metadata;
+  GrpcMetadata(this.metadata);
+
+  @override
+  String toString() => 'gRPC Metadata ($metadata)';
+}
+
+class GrpcData extends GrpcMessage {
+  final List<int> data;
+  final bool isCompressed;
+  GrpcData(this.data, {this.isCompressed});
+
+  @override
+  String toString() => 'gRPC Data (${data.length} bytes)';
+}
+
+StreamTransformer<GrpcMessage, GrpcMessage> grpcDecompressor() =>
+    new StreamTransformer<GrpcMessage, GrpcMessage>.fromHandlers(
+        handleData: (GrpcMessage value, EventSink<GrpcMessage> sink) {
+      if (value is GrpcData) {
+        if (value.isCompressed) {
+          // TODO(dart-lang/grpc-dart#6): Actually handle decompression.
+          sink.add(new GrpcData(value.data, isCompressed: false));
+          return;
+        }
+      }
+      sink.add(value);
+    });
 
 class GrpcHttpEncoder extends Converter<GrpcMessage, StreamMessage> {
   @override
@@ -37,12 +68,22 @@ class GrpcHttpEncoder extends Converter<GrpcMessage, StreamMessage> {
     }
     throw new GrpcError.internal('Unexpected message type');
   }
+
+  static List<int> frame(List<int> payload) {
+    final payloadLength = payload.length;
+    final bytes = new Uint8List(payloadLength + 5);
+    final header = bytes.buffer.asByteData(0, 5);
+    header.setUint8(0, 0); // TODO(dart-lang/grpc-dart#6): Handle compression
+    header.setUint32(1, payloadLength);
+    bytes.setRange(5, bytes.length, payload);
+    return bytes;
+  }
 }
 
 class GrpcHttpDecoder extends Converter<StreamMessage, GrpcMessage> {
   @override
   GrpcMessage convert(StreamMessage input) {
-    final sink = new GrpcMessageSink();
+    final sink = new _GrpcMessageSink();
     startChunkedConversion(sink)
       ..add(input)
       ..close();
@@ -140,5 +181,24 @@ class _GrpcMessageConversionSink extends ChunkedConversionSink<StreamMessage> {
       throw new GrpcError.unavailable('Closed in non-idle state');
     }
     _out.close();
+  }
+}
+
+class _GrpcMessageSink extends Sink<GrpcMessage> {
+  GrpcMessage message;
+
+  @override
+  void add(GrpcMessage data) {
+    if (message != null) {
+      throw 'Too many messages received!';
+    }
+    message = data;
+  }
+
+  @override
+  void close() {
+    if (message == null) {
+      throw 'No messages received!';
+    }
   }
 }
