+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: grpc
 description: Dart implementation of gRPC, a high performance, open-source universal RPC framework.
-version: 1.1.0
+version: 1.0.1
 author: Dart Team <misc@dartlang.org>
 homepage: https://github.com/dart-lang/grpc-dart
 
@@ -8,14 +8,12 @@ environment:
   sdk: '>=2.0.0 <3.0.0'
 
 dependencies:
+  async: '>=1.13.3 <3.0.0'
   googleapis_auth: ^0.2.5+3
   meta: ^1.0.5
   http: '>=0.11.3+17 <0.13.0'
   http2: '>=0.1.7 <2.0.0'
 
 dev_dependencies:
-  build_runner: ^0.10.0
-  build_test: ^0.10.3
-  build_web_compilers: ^0.4.3
   mockito: ^4.0.0
   test: ^1.5.0
