+++ test/client_test.dart
@@ -16,11 +16,11 @@
 import 'dart:async';
 
 import 'package:grpc/grpc.dart';
-import 'package:grpc/src/shared/message.dart';
+import 'package:http2/transport.dart';
 import 'package:test/test.dart';
 
-import '../src/client_utils.dart';
-import '../src/utils.dart';
+import 'src/client_utils.dart';
+import 'src/utils.dart';
 
 void main() {
   const dummyValue = 0;
@@ -39,8 +39,8 @@ void main() {
     const requestValue = 17;
     const responseValue = 19;
 
-    void handleRequest(List<int> message) {
-      final data = validateClientDataMessage(message);
+    void handleRequest(StreamMessage message) {
+      final data = validateDataMessage(message);
       expect(mockDecode(data.data), requestValue);
 
       harness
@@ -63,8 +63,8 @@ void main() {
 
     var index = 0;
 
-    void handleRequest(List<int> message) {
-      final data = validateClientDataMessage(message);
+    void handleRequest(StreamMessage message) {
+      final data = validateDataMessage(message);
       expect(mockDecode(data.data), requests[index++]);
     }
 
@@ -89,8 +89,8 @@ void main() {
     const request = 4;
     const responses = const [3, 17, 9];
 
-    void handleRequest(List<int> message) {
-      final data = validateClientDataMessage(message);
+    void handleRequest(StreamMessage message) {
+      final data = validateDataMessage(message);
       expect(mockDecode(data.data), request);
 
       harness.sendResponseHeader();
@@ -112,8 +112,8 @@ void main() {
 
     var index = 0;
 
-    void handleRequest(List<int> message) {
-      final data = validateClientDataMessage(message);
+    void handleRequest(StreamMessage message) {
+      final data = validateDataMessage(message);
       expect(mockDecode(data.data), requests[index]);
 
       if (index == 0) {
@@ -244,11 +244,10 @@ void main() {
     const customStatusMessage = 'Custom message';
 
     void handleRequest(_) {
-      final headers = <String, String>{
-        'grpc-status': '$customStatusCode',
-        'grpc-message': customStatusMessage
-      };
-      harness.toClient.add(new GrpcMetadata(headers));
+      harness.toClient.add(new HeadersStreamMessage([
+        new Header.ascii('grpc-status', '$customStatusCode'),
+        new Header.ascii('grpc-message', customStatusMessage)
+      ], endStream: true));
       harness.toClient.close();
     }
 
@@ -313,8 +312,8 @@ void main() {
     );
   });
 
-  Future<Null> makeUnaryCall() async {
-    void handleRequest(List<int> message) {
+  Future<void> makeUnaryCall() async {
+    void handleRequest(StreamMessage message) {
       harness
         ..sendResponseHeader()
         ..sendResponseValue(1)
