+++ test/src/client_utils.dart
@@ -15,7 +15,8 @@
 
 import 'dart:async';
 
-import 'package:grpc/src/shared/message.dart';
+import 'package:grpc/src/shared/streams.dart';
+import 'package:http2/transport.dart';
 import 'package:test/test.dart';
 import 'package:mockito/mockito.dart';
 
@@ -23,41 +24,29 @@ import 'package:grpc/grpc.dart';
 
 import 'utils.dart';
 
-typedef void ClientTestMessageHandler(List<int> message);
+class MockTransport extends Mock implements ClientTransportConnection {}
 
-GrpcData validateClientDataMessage(List<int> message) {
-  final decoded = new GrpcData(message);
-
-  expect(decoded, new TypeMatcher<GrpcData>());
-  return decoded;
-}
-
-class MockTransport extends Mock implements Transport {}
-
-class MockStream extends Mock implements GrpcTransportStream {}
+class MockStream extends Mock implements ClientTransportStream {}
 
 class FakeConnection extends ClientConnection {
+  final ClientTransportConnection transport;
+
   var connectionError;
 
-  FakeConnection._(String host, Transport transport, ChannelOptions options,
-      ConnectTransport connectTransport)
-      : super(host, 443, options, connectTransport);
+  FakeConnection(String host, this.transport, ChannelOptions options)
+      : super(host, 443, options);
 
-  factory FakeConnection(
-      String host, Transport transport, ChannelOptions options) {
-    FakeConnection f;
-    f = FakeConnection._(host, transport, options, (_, _1, _2) async {
-      if (f.connectionError != null) throw f.connectionError;
-      return transport;
-    });
-    return f;
+  @override
+  Future<ClientTransportConnection> connectTransport() async {
+    if (connectionError != null) throw connectionError;
+    return transport;
   }
 }
 
 Duration testBackoff(Duration lastBackoff) => const Duration(milliseconds: 1);
 
 class FakeChannelOptions implements ChannelOptions {
-  ChannelCredentials credentials = const Http2ChannelCredentials.secure();
+  ChannelCredentials credentials = const ChannelCredentials.secure();
   Duration idleTimeout = const Duration(seconds: 1);
   BackoffStrategy backoffStrategy = testBackoff;
 }
@@ -72,6 +61,8 @@ class FakeChannel extends ClientChannel {
   @override
   Future<ClientConnection> getConnection() async => connection;
 }
+
+typedef ServerMessageHandler = void Function(StreamMessage message);
 
 class TestClient extends Client {
   static final _$unary =
@@ -119,8 +110,8 @@ class ClientHarness {
   FakeChannelOptions channelOptions;
   MockStream stream;
 
-  StreamController<List<int>> fromClient;
-  StreamController<GrpcMessage> toClient;
+  StreamController<StreamMessage> fromClient;
+  StreamController<StreamMessage> toClient;
 
   TestClient client;
 
@@ -132,7 +123,8 @@ class ClientHarness {
     stream = new MockStream();
     fromClient = new StreamController();
     toClient = new StreamController();
-    when(transport.makeRequest(any, any, any)).thenReturn(stream);
+    when(transport.makeRequest(any, endStream: anyNamed('endStream')))
+        .thenReturn(stream);
     when(transport.onActiveStateChanged = captureAny).thenReturn(null);
     when(stream.outgoingMessages).thenReturn(fromClient.sink);
     when(stream.incomingMessages).thenAnswer((_) => toClient.stream);
@@ -144,17 +136,18 @@ class ClientHarness {
     toClient.close();
   }
 
-  void sendResponseHeader({Map<String, String> headers = const {}}) {
-    toClient.add(new GrpcMetadata(headers));
+  void sendResponseHeader({List<Header> headers = const []}) {
+    toClient.add(new HeadersStreamMessage(headers));
   }
 
   void sendResponseValue(int value) {
-    toClient.add(new GrpcData(mockEncode(value)));
+    toClient
+        .add(new DataStreamMessage(GrpcHttpEncoder.frame(mockEncode(value))));
   }
 
   void sendResponseTrailer(
-      {Map<String, String> headers = const {}, bool closeStream = true}) {
-    toClient.add(new GrpcMetadata(headers));
+      {List<Header> headers = const [], bool closeStream = true}) {
+    toClient.add(new HeadersStreamMessage(headers, endStream: true));
     if (closeStream) toClient.close();
   }
 
@@ -171,11 +164,11 @@ class ClientHarness {
       String expectedPath,
       Duration expectedTimeout,
       Map<String, String> expectedCustomHeaders,
-      List<ClientTestMessageHandler> serverHandlers = const [],
+      List<MessageHandler> serverHandlers = const [],
       Function doneHandler,
       bool expectDone = true}) async {
     int serverHandlerIndex = 0;
-    void handleServerMessage(List<int> message) {
+    void handleServerMessage(StreamMessage message) {
       serverHandlers[serverHandlerIndex++](message);
     }
 
@@ -189,17 +182,12 @@ class ClientHarness {
       expect(result, expectedResult);
     }
 
-    final capturedParameters =
-        verify(transport.makeRequest(captureAny, captureAny, captureAny))
-            .captured;
-    if (expectedPath != null) {
-      expect(capturedParameters[0], expectedPath);
-    }
-    expect(capturedParameters[1], expectedTimeout);
-    final Map<String, String> headers = capturedParameters[2];
-    headers?.forEach((key, value) {
-      expect(expectedCustomHeaders[key], value);
-    });
+    final List<Header> capturedHeaders =
+        verify(transport.makeRequest(captureAny)).captured.single;
+    validateRequestHeaders(capturedHeaders,
+        path: expectedPath,
+        timeout: toTimeoutString(expectedTimeout),
+        customHeaders: expectedCustomHeaders);
 
     await clientSubscription.cancel();
   }
@@ -219,7 +207,7 @@ class ClientHarness {
       String expectedPath,
       Duration expectedTimeout,
       Map<String, String> expectedCustomHeaders,
-      List<ClientTestMessageHandler> serverHandlers = const [],
+      List<MessageHandler> serverHandlers = const [],
       bool expectDone = true}) async {
     return runTest(
       clientCall: expectThrows(clientCall, expectedException),
