+++ test/src/utils.dart
@@ -16,7 +16,6 @@
 import 'dart:convert';
 
 import 'package:grpc/src/shared/streams.dart';
-import 'package:grpc/src/shared/message.dart';
 import 'package:http2/transport.dart';
 import 'package:test/test.dart';
 
@@ -30,25 +29,26 @@ Map<String, String> headersToMap(List<Header> headers) =>
     new Map.fromIterable(headers,
         key: (h) => ascii.decode(h.name), value: (h) => ascii.decode(h.value));
 
-void validateRequestHeaders(Map<String, String> headers,
+void validateRequestHeaders(List<Header> headers,
     {String path,
     String authority = 'test',
     String timeout,
     Map<String, String> customHeaders}) {
-  expect(headers[':method'], 'POST');
-  expect(headers[':scheme'], 'https');
+  final headerMap = headersToMap(headers);
+  expect(headerMap[':method'], 'POST');
+  expect(headerMap[':scheme'], 'https');
   if (path != null) {
-    expect(headers[':path'], path);
+    expect(headerMap[':path'], path);
   }
-  expect(headers[':authority'], authority);
-  expect(headers['grpc-timeout'], timeout);
-  expect(headers['content-type'], 'application/grpc');
-  expect(headers['te'], 'trailers');
-  expect(headers['grpc-accept-encoding'], 'identity');
-  expect(headers['user-agent'], startsWith('dart-grpc/'));
+  expect(headerMap[':authority'], authority);
+  expect(headerMap['grpc-timeout'], timeout);
+  expect(headerMap['content-type'], 'application/grpc');
+  expect(headerMap['te'], 'trailers');
+  expect(headerMap['grpc-accept-encoding'], 'identity');
+  expect(headerMap['user-agent'], startsWith('dart-grpc/'));
 
   customHeaders?.forEach((key, value) {
-    expect(headers[key], value);
+    expect(headerMap[key], value);
   });
 }
